+++ lib/presentation/screens/Client/home/view/client_feedback.dart
@@ -58,11 +58,11 @@ class ClientFeedbackView extends StatelessWidget {
                   ),
                   SizedBox(height: AppSize.s12),
                   StarRating(
-                    rating: cubit.rating,
+                    rating: cubit.rating.toDouble(),
                     starCount: 5,
                     size: AppSize.s35,
                     onRatingChanged: (rating) {
-                      cubit.changeRating(rating);
+                      cubit.changeRating(rating.toInt());
                       cubit.changeFeedbackValidation();
                     },
                   ),
@@ -90,7 +90,7 @@ class ClientFeedbackView extends StatelessWidget {
                           onPressed: cubit.feedbackValid == false
                               ? null
                               : () async {
-                                  await cubit.sendFeedback(context,toUser);
+                                  await cubit.sendFeedback(context, toUser);
                                 },
                           width: double.infinity,
                         ),
